lwc-events: test case for event based on OTEL trace (#1816)
Adding a test case to show how an event could be modelled over
a trace using the OTEL model. The event has some basic structure
for the trace to allow efficient access to the root and parent
spans to support querying trace shape from an event. Child is not
supported directly, but can be rewritten to query based on the
parent if needed.

To avoid recomputing some of the trace structure for each span
while processing, it provides a `foreach` to iterate over the
spans for the trace when processing.