lwcapi: disable query index cache (#1841)
Disables the query index cache used to find the expressions
for a given cluster. There is a subtle race condition that
can lead to stale data being cached and the expression not
showing up for the cluster. Disable for now and revisit when
there is a fix for the underlying issue.